update zerodev example
