+++ examples/with-zerodev-aa/src/util.ts
@@ -10,7 +10,7 @@ export function assertEqual<T>(left: T, right: T) {
 
 export function refineNonNull<T>(
   input: T | null | undefined,
-  errorMessage?: string
+  errorMessage?: string,
 ): T {
   if (input == null) {
     throw new Error(errorMessage ?? `Unexpected ${JSON.stringify(input)}`);
